% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_mm_to_duration.R
\name{conv_mm_to_duration}
\alias{conv_mm_to_duration}
\title{conv_mm_to_duration}
\usage{
conv_mm_to_duration(mmOfWater)
}
\arguments{
\item{mmOfWater}{depth of desired water column}
}
\value{
Seconds to open valve based on above assumptions
}
\description{
Converts mm of water to seconds to open valves
}
\details{
emitters are spaced every 254 mm and emit 946353 mm^3 per hour
}
