% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_waterByZone.R
\name{create_waterByZone}
\alias{create_waterByZone}
\title{create_waterByZone}
\usage{
create_waterByZone()
}
\value{
void
}
\description{
Set up an initial matrix to track weather and watering
}
\details{
creates waterByZone.RDS on local file system
}
