% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howMuchToWater.R
\name{howMuchToWater}
\alias{howMuchToWater}
\title{How Much To Water each day}
\usage{
howMuchToWater(waterByZone, yearDay)
}
\arguments{
\item{waterByZone}{matrix showing rainfall, needed and watered}

\item{yearDay}{day of the year. January 1 = 1}
}
\value{
updated WaterByZone
}
\description{
compares rainfall from yesterday, today, and tomorrow
   minus evapotranspiration
   against today's needed amount of water
}
