% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_heartbeat.R
\name{send_heartbeat}
\alias{send_heartbeat}
\title{Send_heartbeat}
\usage{
send_heartbeat(waterByZone)
}
\arguments{
\item{waterByZone}{A copy of the waterByZone matrix}
}
\value{
void
}
\description{
Let the web page know we are alive and working
}
