% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_waterbyzone.R
\name{update_waterbyzone}
\alias{update_waterbyzone}
\title{update waterByZone with current forecast}
\usage{
update_waterbyzone(waterByZone)
}
\arguments{
\item{waterByZone}{a matrix with zone requests and rain forecast}
}
\value{
updated version of the waterByZone matrix
}
\description{
update waterByZone with current forecast
}
