% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_mm_to_duration.R
\name{conv_mm_to_duration}
\alias{conv_mm_to_duration}
\title{conv_mm_to_duration}
\usage{
conv_mm_to_duration(mmOfWater)
}
\arguments{
\item{mmOfWater}{depth of desired water column}
}
\value{
Seconds to open valve based on above assumptions
}
\description{
Converts mm of water to seconds to open valves
}
\details{
emitters are spaced every 254 mm and emit 946 ml per hour
1 mm of water with emitters spaced every 254 mm = 64.51 ml of water
it will take one emitter 252 seconds to emit 64.51 ml of water
}
