% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigate.R
\name{irrigate}
\alias{irrigate}
\title{irrigate}
\usage{
irrigate(state, zone)
}
\arguments{
\item{state}{Logical. TRUE = on, FALSE = off}

\item{zone}{numerical. 1 = front yard, 2 = back yard}
}
\value{
nothing
}
\description{
Turns the irrigation system on and off
}
