% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_heartbeat.R
\name{send_heartbeat}
\alias{send_heartbeat}
\title{Send_heartbeat}
\usage{
send_heartbeat()
}
\arguments{
\item{iam_to_send}{ip address of raspberry pi}
}
\value{
void
}
\description{
Let the web page know we are alive and working
}
