% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_waterbyzone.R
\name{update_waterbyzone}
\alias{update_waterbyzone}
\title{update waterByZone with current forecast}
\usage{
update_waterbyzone(waterByZone, yearDay)
}
\arguments{
\item{waterByZone}{a matrix with zone requests and rain forecast}

\item{yearDay}{the day of the year. January 1 = 0}
}
\value{
updated version of the waterByZone matrix
}
\description{
update waterByZone with current forecast
}
